move remaining armory files to charmory - refactor references
