+++ examples/src/armory/examples/image_classification/food101/food101_pgd.py
@@ -1,19 +1,19 @@
 from pprint import pprint
 import sys
 
+import armory.baseline_models.pytorch.food101
+import armory.data.datasets
+import armory.version
 import art.attacks.evasion
 from jatic_toolbox import load_dataset as load_jatic_dataset
 import torch
 import torchmetrics.classification
 from torchvision import transforms as T
 
-import armory.baseline_models.pytorch.food101
-import armory.data.datasets
-from armory.metrics.compute import BasicProfiler
-import armory.version
 from charmory.data import ArmoryDataLoader
 from charmory.engine import EvaluationEngine
 from charmory.evaluation import Dataset, Evaluation, Metric, Model, SysConfig
+from charmory.metrics.compute import BasicProfiler
 from charmory.metrics.perturbation import PerturbationNormMetric
 from charmory.perturbation import ArtEvasionAttack
 from charmory.tasks.image_classification import ImageClassificationTask
