+++ examples/src/armory/examples/image_classification/imagenet1k_resnet34_pgd.py
@@ -4,6 +4,7 @@ Example programmatic entrypoint for scenario execution
 from pprint import pprint
 import sys
 
+import armory.version
 import art.attacks.evasion
 from art.estimators.classification import PyTorchClassifier
 from jatic_toolbox import __version__ as jatic_version
@@ -13,12 +14,11 @@ import torch
 import torch.nn as nn
 import torchmetrics.classification
 
-from armory.metrics.compute import BasicProfiler
-import armory.version
 from charmory.data import ArmoryDataLoader
 from charmory.engine import EvaluationEngine
 from charmory.evaluation import Dataset, Evaluation, Metric, Model
 from charmory.experimental.transforms import create_image_classification_transform
+from charmory.metrics.compute import BasicProfiler
 from charmory.metrics.perturbation import PerturbationNormMetric
 from charmory.model.image_classification import JaticImageClassificationModel
 from charmory.perturbation import ArtEvasionAttack
