+++ examples/src/armory/examples/image_classification/kaludi_food_category_pgd.py
@@ -1,5 +1,6 @@
 from pprint import pprint
 
+from armory.examples.utils.args import create_parser
 import art.attacks.evasion
 from art.estimators.classification import PyTorchClassifier
 import jatic_toolbox
@@ -9,11 +10,10 @@ import torch.nn
 import torchmetrics.classification
 from transformers.image_utils import infer_channel_dimension_format
 
-from armory.examples.utils.args import create_parser
-from armory.metrics.compute import BasicProfiler
 from charmory.data import ArmoryDataLoader
 from charmory.engine import EvaluationEngine
 from charmory.evaluation import Dataset, Evaluation, Metric, Model
+from charmory.metrics.compute import BasicProfiler
 from charmory.metrics.perturbation import PerturbationNormMetric
 from charmory.model.image_classification import JaticImageClassificationModel
 from charmory.perturbation import ArtEvasionAttack
