+++ library/src/charmory/evaluation.py
@@ -8,7 +8,7 @@ from art.estimators import BaseEstimator
 from torch.utils.data.dataloader import DataLoader
 from torchmetrics.metric import Metric as TorchMetric
 
-from armory.metrics.compute import NullProfiler, Profiler
+from charmory.metrics.compute import NullProfiler, Profiler
 from charmory.perturbation import Perturbation
